Module measureme::file_header

All binary files generated by measureme have a simple file header that consists of a 4 byte file magic string and a 4 byte little-endian version number.

Constants

CURRENT_FILE_FORMAT_VERSION

FILE_EXTENSION

FILE_HEADER_SIZE

The size of the file header in bytes. Note that functions in this module rely on this size to be 8.

FILE_MAGIC_EVENT_STREAM

FILE_MAGIC_STRINGTABLE_DATA

FILE_MAGIC_STRINGTABLE_INDEX

FILE_MAGIC_TOP_LEVEL

Functions

strip_file_header

verify_file_header

write_file_header