Function measureme::split_streams

pub fn split_streams(paged_data: &[u8]) -> FxHashMap<PageTag, Vec<u8>>

This function reconstructs the individual data streams from their paged version.

For example, if E denotes the page header of an events page, S denotes the header of a string data page, and lower case letters denote page contents then a paged stream could look like:

s = Eabcd_Sopq_Eef_Eghi_Srst

and split_streams would result in the following set of streams:

split_streams(s) = {
    events: [abcdefghi],
    string_data: [opqrst],
}