Struct measureme::RawEvent

#[repr(C)]
pub struct RawEvent {
    pub event_kind: StringId,
    pub event_id: EventId,
    pub thread_id: u32,
    pub payload1_lower: u32,
    pub payload2_lower: u32,
    pub payloads_upper: u32,
}

RawEvent is how events are stored on-disk. If you change this struct, make sure that you increment file_header::CURRENT_FILE_FORMAT_VERSION.

Fields
event_kind: StringIdevent_id: EventIdthread_id: u32payload1_lower: u32payload2_lower: u32payloads_upper: u32
Implementations

impl RawEvent

pub fn new_interval( event_kind: StringId, event_id: EventId, thread_id: u32, start: u64, end: u64 ) -> Self

pub fn new_instant( event_kind: StringId, event_id: EventId, thread_id: u32, instant: u64 ) -> Self

pub fn new_integer( event_kind: StringId, event_id: EventId, thread_id: u32, value: u64 ) -> Self

pub fn start_value(&self) -> u64

The start value assuming self is an interval

pub fn end_value(&self) -> u64

The end value assuming self is an interval

pub fn value(&self) -> u64

The value assuming self is an interval or integer.

pub fn is_instant(&self) -> bool

pub fn is_integer(&self) -> bool

pub fn serialize(&self, bytes: &mut [u8])

pub fn deserialize(bytes: &[u8]) -> RawEvent